Sprint 2 (#591 E5c): split AudienceAdminAudience into controller + renderer
Extracts the HTML-rendering methods (render_list, render_row_recursive,
render_form, render_custom_fields_section, render_custom_field_row,
render_members) into AudienceAdminAudienceRenderer (static, explicit params —
the menu slug is now a leading arg). The controller keeps render_page as
orchestrator, handle_actions, and the constructor. Controller 899 → 196 lines.

Markup byte-identical (only internal $this-> → self:: call-site rewrites);
DB/seed calls, AdminUI::render_toggle, escaping, nonces unchanged. Renderer
carries @covers + setUp preload (verified 409/454). Behavior-preserving —
Audience suite green (654 tests); PHPStan L8 + WPCS clean.
